-- numerical constants for zero and one extended and modularized api.c/.h 	- extended getNumberOfHexDigits to care for numerical constants for 1/0 		- by using the new fcn 			- isNumericalConstantZero 			- isNumericalConstantOne 	- added bool fcn isNumericArgument - consequences for 	api_debug.c one_wire.c and relay.c 		- replacing simple checks on getNumberOfHexDigits == 0 by isNumericArgument 		- getUnsignedNumericValueFromParameterIndex instead of local strtoul conversion

+++ relay.c
@@ -142,8 +142,8 @@ void relayThreshold(struct uartStruct *ptr_uartStruct)
 void relayThresholdMiscSubCommands( struct uartStruct *ptr_uartStruct, int16_t subCommandIndex )
 {
    uint32_t status = 0;
-   uint32_t value = 0;
-   uint32_t value2 = 0;
+   uint64_t value = 0;
+   uint64_t value2 = 0;
    uint8_t channel = 0;
    uint8_t recursive = FALSE;
